sdk-python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ntt-client-auth"
version = "0.1.0"
description = "Python client SDK for Centralized Authentication Service"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "pydantic",
    "PyJWT[crypto]>=2.8.0",
    "cryptography>=42.0.0",
    "httpx>=0.27.0"
]

[project.optional-dependencies]
dev = [
    "pytest"
]
